# openapi.yaml
openapi: 3.0.3
info:
  title: RentCar API — Sistema de Alquiler de Autos
  description: |
    API REST para el sistema de alquiler de autos. Arquitectura Clean Architecture
    con módulos de vehículos, reservas, alquileres, devoluciones, pagos y facturación.

    ## Flujo principal
    1. El cliente busca vehículos disponibles (`GET /vehiculos/search`)
    2. Crea una reserva (`POST /reservas`)
    3. Realiza el pago (`POST /pagos`)
    4. El admin inicia el alquiler (`POST /alquileres`)
    5. Al terminar, el admin registra la devolución (`POST /devoluciones`)
    6. Se genera la factura (`POST /facturas`)

    ## Autenticación
    JWT Bearer token. Obtener en `POST /auth/login`.

    ## Preparado para integración futura (Reto 2)
    Booking Prototipo usará: `GET /vehiculos/search`, `POST /reservas`,
    `PATCH /reservas/{id}/cancel`, `POST /pagos`.

    Eventos outbox modelados: `RESERVA_CREADA`, `RESERVA_CANCELADA`,
    `PAGO_REGISTRADO`, `ALQUILER_INICIADO`, `VEHICULO_DEVUELTO`, `FACTURA_GENERADA`.
  version: 1.0.0
  contact:
    name: RentCar Ecuador
    email: dev@rentcar.ec
servers:
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
  - url: https://rentcar-api.render.com/api/v1
    description: Producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateVehiculoRequest:
      type: object
      required: [placa, agenciaId, modeloId, categoriaId, tipoCombustibleId, tipoTransmisionId, color, anio, precioDia]
      properties:
        placa:             { type: string }
        agenciaId:         { type: string, format: uuid }
        modeloId:          { type: string, format: uuid }
        categoriaId:       { type: string, format: uuid }
        tipoCombustibleId: { type: string, format: uuid }
        tipoTransmisionId: { type: string, format: uuid }
        color:             { type: string }
        anio:              { type: integer }
        kilometraje:       { type: integer, default: 0 }
        numeroPasajeros:   { type: integer, default: 5 }
        precioDia:         { type: number }
        imagenUrl:         { type: string }
        descripcion:       { type: string }
    CreateReservaRequest:
      type: object
      required: [vehiculoId, agenciaId, fechaInicio, fechaFin]
      properties:
        vehiculoId:   { type: string, format: uuid }
        agenciaId:    { type: string, format: uuid }
        fechaInicio:  { type: string, format: date, example: "2025-06-01" }
        fechaFin:     { type: string, format: date, example: "2025-06-04" }
        seguroId:     { type: string, format: uuid }
        canalVentaId: { type: string, format: uuid }
        extras:
          type: array
          items:
            type: object
            properties:
              extraId:  { type: string, format: uuid }
              cantidad: { type: integer, default: 1 }
        notas: { type: string }
    CreateAlquilerRequest:
      type: object
      required: [reservaId, kmSalida]
      properties:
        reservaId:     { type: string, format: uuid }
        kmSalida:      { type: integer }
        observaciones: { type: string }
    CreateDevolucionRequest:
      type: object
      required: [alquilerId, kmEntrada, estadoVehiculo]
      properties:
        alquilerId:     { type: string, format: uuid }
        kmEntrada:      { type: integer }
        estadoVehiculo: { type: string, example: "Bueno" }
        cargoExtra:     { type: number, default: 0 }
        observaciones:  { type: string }
    CreatePagoRequest:
      type: object
      required: [reservaId, monto, metodoPago]
      properties:
        reservaId:  { type: string, format: uuid }
        monto:      { type: number }
        metodoPago: { type: string, enum: [EFECTIVO, TARJETA, TRANSFERENCIA] }
        referencia: { type: string }
    CreateFacturaRequest:
      type: object
      required: [reservaId]
      properties:
        reservaId:   { type: string, format: uuid }
        pagoId:      { type: string, format: uuid }
        rucCliente:  { type: string }
        razonSocial: { type: string }
paths:
...
  /vehiculos:
    get:
      tags: [Vehículos]
      summary: Listar vehículos (paginado)
      parameters:
        - { in: query, name: page,  schema: { type: integer, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Lista paginada }
    post:
      tags: [Vehículos]
      summary: Crear vehículo (ADMIN)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVehiculoRequest' }
      responses:
        '201': { description: Vehículo creado }
...
  /reservas:
    get:
      tags: [Reservas]
      summary: Listar todas las reservas (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Lista de reservas }
    post:
      tags: [Reservas]
      summary: Crear reserva
      description: |
        El backend recalcula el precio. Valida solapamiento de fechas.
        Publica evento RESERVA_CREADA en outbox.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReservaRequest' }
      responses:
        '201': { description: Reserva creada }
        '422': { description: No disponible o fechas solapadas }
...
  /alquileres:
    get:
      tags: [Alquileres]
      summary: Listar alquileres (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Lista de alquileres }
    post:
      tags: [Alquileres]
      summary: Iniciar alquiler (ADMIN)
      description: Cambia vehículo a EN_USO, reserva a ACTIVA. Registra Kardex.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAlquilerRequest' }
      responses:
        '201': { description: Alquiler iniciado }
...
  /devoluciones:
    get:
      tags: [Devoluciones]
      summary: Listar devoluciones (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Lista de devoluciones }
    post:
      tags: [Devoluciones]
      summary: Registrar devolución (ADMIN)
      description: Finaliza alquiler, vehículo vuelve a DISPONIBLE. Registra Kardex.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDevolucionRequest' }
      responses:
        '201': { description: Devolución registrada }
...
  /pagos:
    get:
      tags: [Pagos]
      summary: Listar pagos (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Lista de pagos }
    post:
      tags: [Pagos]
      summary: Registrar pago
      description: Valida que el monto cubra el total. Confirma la reserva si estaba PENDIENTE.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePagoRequest' }
      responses:
        '201': { description: Pago registrado }
        '400': { description: Monto insuficiente }
...
  /facturas:
    get:
      tags: [Facturas]
      summary: Listar facturas (ADMIN)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Lista de facturas }
    post:
      tags: [Facturas]
      summary: Generar factura
      description: Calcula subtotal, IVA 15% y total. Genera detalles automáticamente.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateFacturaRequest' }
      responses:
        '201': { description: Factura generada }
...
  /admin/agencias:
    get:  { tags: [Admin], summary: Listar agencias,  security: [{ BearerAuth: [] }], responses: { '200': { description: Lista } } }
    post: { tags: [Admin], summary: Crear agencia,    security: [{ BearerAuth: [] }], requestBody: { content: { application/json: { schema: { type: object } } } }, responses: { '201': { description: Creada } } }
  /admin/agencias/{id}:
    parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
    get:    { tags: [Admin], summary: Detalle agencia,     security: [{ BearerAuth: [] }], responses: { '200': { description: OK } } }
    patch:  { tags: [Admin], summary: Actualizar agencia,  security: [{ BearerAuth: [] }], requestBody: { content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: OK } } }
    delete: { tags: [Admin], summary: Eliminar agencia,    security: [{ BearerAuth: [] }], responses: { '200': { description: OK } } }
  /admin/empresas:
    get:  { tags: [Admin], summary: Listar empresas,  security: [{ BearerAuth: [] }], responses: { '200': { description: Lista } } }
    post: { tags: [Admin], summary: Crear empresa,    security: [{ BearerAuth: [] }], requestBody: { content: { application/json: { schema: { type: object } } } }, responses: { '201': { description: Creada } } }
  /admin/empresas/{id}:
    parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
    get:    { tags: [Admin], summary: Detalle empresa,    security: [{ BearerAuth: [] }], responses: { '200': { description: OK } } }
    patch:  { tags: [Admin], summary: Actualizar empresa, security: [{ BearerAuth: [] }], requestBody: { content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: OK } } }
    delete: { tags: [Admin], summary: Eliminar empresa,   security: [{ BearerAuth: [] }], responses: { '200': { description: OK } } }
...